Add support for `ignored` config (#371)
* add ignored to the config object

* exclude ignored packages in release

* throw for mixed changesets

* Do not delete changeset files for ignored packages

* fix a bug where an unchanged package gets a patch release

* do not bump ignored packages in prereleases

* use Array.prototype.includes

* add input validation

* filter out ignored packages in flattenReleases()

* address comments

* correct typo

* add --ignore to cli

* rename ignored to ignore

* Update error message format

* add config tests

* add tests for apply-release-plan

* Add tests for assemble-release-plan

* address comment

* add tests for the version command

* add a test for the pre exit mode fix

* show all missing packages when validating --ignore flag

* fix lint errors

* Throws when --ignore flag and ignore in config are both present

* handle --ignore and ignore in config before calling version()

* change getDependents() to take a single option argument

* pass an option to determineDependents()

* add a test for incrementVersion()

* update test case description

* Update packages/assemble-release-plan/src/index.ts


* run prettier

* Require listing all dependents in the ignore array

* fix lint

* sort deps

* update test message

* add tests for version for cli

* fix tests + prettier

* refactor to make test easier

* run prettier

* Create rich-melons-press.md

* remove code for debugging

* add integration tests

* fix test case

* remove line breaks

* remove line break again

* run prettier

* Create proud-tips-allow.md

* Create proud-tips-allow.md

* Update rich-melons-press.md

* Update proud-tips-allow.md

* Update .changeset/rich-melons-press.md


* Update packages/cli/src/run.ts


* Update packages/config/src/index.ts


* Update packages/config/src/index.ts


* Update .changeset/proud-tips-allow.md


* Update packages/cli/src/run.ts


* address comments

* fix lint issue

* add missing bracket

* fix merge issue

+++ packages/apply-release-plan/src/index.ts
@@ -135,8 +135,19 @@ export default async function applyReleasePlan(
         let changesetPath = path.resolve(changesetFolder, `${changeset.id}.md`);
         let changesetFolderPath = path.resolve(changesetFolder, changeset.id);
         if (await fs.pathExists(changesetPath)) {
-          touchedFiles.push(changesetPath);
-          await fs.remove(changesetPath);
+          // DO NOT remove changeset for ignored packages
+          // Mixed changeset that contains both ignored packages and not ignored packages are disallowed
+          // At this point, we know there is no such changeset, because otherwise the program would've already failed,
+          // so we just check if any ignored package exists in this changeset, and only remove it if none exists
+          // Ignored list is added in v2, so we don't need to do it for v1 changesets
+          if (
+            !changeset.releases.find(release =>
+              config.ignore.includes(release.name)
+            )
+          ) {
+            touchedFiles.push(changesetPath);
+            await fs.remove(changesetPath);
+          }
           // TO REMOVE LOGIC - this works to remove v1 changesets. We should be removed in the future
         } else if (await fs.pathExists(changesetFolderPath)) {
           touchedFiles.push(changesetFolderPath);
